Testing with terraform
This required a number of make additions as well as some external
tooling to work with the https://github.com/erikh/tftest library. This
library is still a work in progress and needs some shaping up, so you
will notice that if these tests fail, that you will still need to
manually reap resources in some, but not all instances.

notably a terraform configuration file and plugins dir were created and
conifgured as a part of the `make test` process, and referred to in a
TestMain function in `main_test.go`. The point of this is to couple the
tests to the build process, which is a requirement in this situation
with how the plugins are loaded. This establishes the `.tfdata`
directory for plugins. Our plugin is unofficially named
`hashicorp/zerotier` and is overridden in the rc file generated by
`mktfrc.sh`, another part of the `make test` process.

The `test-token.txt` can be used to locally store or mount as secret a
zerotier token without committing it to the repository, I felt it might
be useful for future development later.

I will attempt to document more of this in the next commit or two.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>