schemawrap: a new approach to synchronizing state between terraform and zerotier
This allows us to keep the synchronization code small and rely on struct
properties with hooks, similar in function to how terraform works
itself. However, we have added additional functionality for the
automatic conversion of types and a very basic validations and
conversions system that is largely write-through.

It uses global variables as a staging ground, but I think that's ok at
this point.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>